RavenDB-23453 Moves GrowableBitArray into Sparrow project.
