+++ test/SlowTests/Corax/GrowableBitArrayTests.cs
@@ -6,6 +6,7 @@ using Corax.Utils;
 using FastTests;
 using FastTests.Voron.FixedSize;
 using Sparrow.Server;
+using Sparrow.Server.Collections;
 using Sparrow.Threading;
 using Tests.Infrastructure;
 using Xunit;
