+++ test/StressTests/Corax/GrowableBitArrayTests.cs
@@ -2,6 +2,7 @@
 using Corax.Utils;
 using FastTests;
 using Sparrow.Server;
+using Sparrow.Server.Collections;
 using Sparrow.Threading;
 using Tests.Infrastructure;
 using Xunit;
